+++ src/dense_neural_net.f08
@@ -110,7 +110,7 @@ subroutine dnn_add_layer(this, out_nodes, activation, drop_rate)
     class(DenseNN)             :: this
     integer, intent(in)        :: out_nodes
     character(*), intent(in)   :: activation
-    real, intent(in)           :: drop_rate
+    real(kind=8), intent(in)   :: drop_rate
     class(DenseLayer), pointer :: new_layer
     integer                    :: in_nodes
 
@@ -162,9 +162,9 @@ end subroutine
 ! alters :: this DenseNN's DenseLayers' z's and a's calculated
 !-------------------------------------------------------------------------------
 subroutine dnn_forw_prop(this, input, is_train)
-    class(DenseNN)      :: this
-    real, intent(in)    :: input(:,:)
-    logical, intent(in) :: is_train
+    class(DenseNN)           :: this
+    real(kind=8), intent(in) :: input(:,:)
+    logical, intent(in)      :: is_train
 
     call this%first_hid%dense_forw_prop(input, is_train)
 
@@ -184,7 +184,7 @@ end subroutine
 !-------------------------------------------------------------------------------
 subroutine dnn_out_delta(this, labels, loss)
     class(DenseNN)           :: this
-    real, intent(in)         :: labels(:,:)
+    real(kind=8), intent(in) :: labels(:,:)
     character(*), intent(in) :: loss
 
     select case (loss)
@@ -240,7 +240,7 @@ end subroutine
 !-------------------------------------------------------------------------------
 subroutine dnn_back_prop(this, labels, loss)
     class(DenseNN)           :: this
-    real, intent(in)         :: labels(:,:)
+    real(kind=8), intent(in) :: labels(:,:)
     character(*), intent(in) :: loss
 
     call this%dnn_out_delta(labels, loss)
@@ -261,9 +261,9 @@ end subroutine
 ! alters ::   this DenseNN's weights and biases adjusted to minimize loss
 !-------------------------------------------------------------------------------
 subroutine dnn_update(this, input, learn_rate, is_train)
-    class(DenseNN)      :: this
-    real, intent(in)    :: input(:,:), learn_rate
-    logical, intent(in) :: is_train
+    class(DenseNN)           :: this
+    real(kind=8), intent(in) :: input(:,:), learn_rate
+    logical, intent(in)      :: is_train
 
     ! first_hid's a(l-1) is input batch
     call this%first_hid%dense_update(input, learn_rate, is_train)
